✨ Feature: 내 정보 수정을 위한 스키마 정의,  휴대폰 인증로직 첨부하여 인증기능 구현 (#229)
